pyproject.toml
[tool.poetry]
name = "mobile-app"
version = "0.1.0"
description = ""
authors = ["Viktoriia Lavrova <viktoriia.lav2908@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.2"
allure-pytest = "^2.13.5"
selene = "2.0.0rc9"
python-dotenv = "^1.0.1"
appium-python-client = "^4.1.0"
requests = "^2.32.3"
pydantic-settings = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
